backend/requirements.txt
# FastAPI and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0

# Database and authentication
supabase>=2.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# OCR and document processing
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python>=4.8.0
PyMuPDF>=1.23.0
pdfplumber>=0.10.0

# NLP and AI
spacy[cuda12x]>=3.7.0
transformers>=4.35.0
torch>=2.1.0
numpy>=2.0.0  # Use numpy 2.x for spacy 3.8+ compatibility

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
loguru>=0.7.0
celery>=5.3.0
redis>=5.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
pydantic-settings>=2.2.0